dashboard/requirements.txt
Flask>=2.0
flask-basicauth>=0.2.0
gunicorn>=20.0
watchdog>=2.1
python-dotenv>=1.0
